Utils of GUI moved to a separate module, tests fixed.
